[VS-7291] lookup services usage refactored to use API generated service and model
