+++ vsu-app/ClientApp/src/app/shared/components/recipient-details/recipient-details.component.ts
@@ -13,8 +13,8 @@ import { MomentDateAdapter } from '@angular/material-moment-adapter';
 import { DateAdapter, MAT_DATE_FORMATS, MAT_DATE_LOCALE } from '@angular/material/core';
 import { noop, Observable, Observer, of, throwError } from 'rxjs';
 import { map, switchMap, tap } from 'rxjs/operators';
+import { LookupService } from 'src/api/lookup/lookup.service';
 import { config } from '../../../../config';
-import { LookupService } from '../../../services/lookup.service';
 import { ApplicationType, MY_FORMATS } from '../../enums-list';
 import { FormBase } from '../../form-base';
 import { CitiesSearchResponse, iCity, iLookupData } from '../../interfaces/lookup-data.interface';
@@ -99,20 +99,27 @@ export class RecipientDetailsComponent extends FormBase implements OnInit {
       switchMap((query: string) => {
         if (query) {
           let searchVal = this.form.get('victimServiceWorker')['controls'][0].get('city').value.toString();
-          return this.lookupService.searchCities(this.countryVal, this.provinceVal, searchVal).pipe(
-            map((data: CitiesSearchResponse) => {
-              if (data && data.CityCollection) {
-                data.CityCollection.sort((a, b) => a.vsd_name.localeCompare(b.vsd_name));
-                return data.CityCollection;
-              } else return [];
-            }),
-            tap(
-              () => noop,
-              (err) => {
-                this.errorMessage = (err && err.message) || 'Something goes wrong';
-              }
-            )
-          );
+          return this.lookupService
+            .getApiLookupCitiesSearch<any>({
+              country: this.countryVal,
+              province: this.provinceVal,
+              searchVal,
+              limit: 15
+            })
+            .pipe(
+              map((data: CitiesSearchResponse) => {
+                if (data && data.CityCollection) {
+                  data.CityCollection.sort((a, b) => a.vsd_name.localeCompare(b.vsd_name));
+                  return data.CityCollection;
+                } else return [];
+              }),
+              tap(
+                () => noop,
+                (err) => {
+                  this.errorMessage = (err && err.message) || 'Something goes wrong';
+                }
+              )
+            );
         }
         return of([]);
       })
