+++ vsu-app/ClientApp/src/app/shared/components/travel-overview/travel-overview.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { ControlContainer, FormArray, FormBuilder, FormGroup } from '@angular/forms';
-import { LookupService } from '../../../services/lookup.service';
+import { LookupService } from 'src/api/lookup/lookup.service';
 import { NotificationQueueService } from '../../../services/notification-queue.service';
 import { ApplicationType } from '../../enums-list';
 import { FormBase } from '../../form-base';
@@ -39,7 +39,7 @@ export class TravelOverviewComponent extends FormBase implements OnInit {
       this.offenceList = this.lookupData.offences;
       this.populateOffences();
     } else {
-      this.lookupService.getOffences().subscribe(
+      this.lookupService.getApiLookupOffences<any>().subscribe(
         (res) => {
           this.lookupData.offences = res.value;
           if (this.lookupData.offences) {
