+++ vsu-app/ClientApp/src/app/victim-travel-fund-application/vtf-application.component.ts
@@ -5,7 +5,7 @@ import { MatStepper } from '@angular/material/stepper';
 import { Title } from '@angular/platform-browser';
 import { Router } from '@angular/router';
 import { ApplicationService } from 'src/api/application/application.service';
-import { LookupService } from '../services/lookup.service';
+import { LookupService } from 'src/api/lookup/lookup.service';
 import { NotificationQueueService } from '../services/notification-queue.service';
 import { ApplicantInfoHelper } from '../shared/components/applicant-information/applicant-information.helper';
 import { AuthInfoHelper } from '../shared/components/authorization/authorization.helper';
@@ -93,7 +93,7 @@ export class VictimTravelFundApplicationComponent extends FormBase implements On
 
     promise_array.push(
       new Promise<void>((resolve, reject) => {
-        this.lookupService.getCountries().subscribe(
+        this.lookupService.getApiLookupCountries<any>().subscribe(
           (res) => {
             this.lookupData.countries = res.value;
             if (this.lookupData.countries) {
@@ -110,7 +110,7 @@ export class VictimTravelFundApplicationComponent extends FormBase implements On
 
     promise_array.push(
       new Promise<void>((resolve, reject) => {
-        this.lookupService.getProvinces().subscribe(
+        this.lookupService.getApiLookupProvinces<any>().subscribe(
           (res) => {
             this.lookupData.provinces = res.value;
             if (this.lookupData.provinces) {
@@ -127,7 +127,7 @@ export class VictimTravelFundApplicationComponent extends FormBase implements On
 
     promise_array.push(
       new Promise<void>((resolve, reject) => {
-        this.lookupService.getOffences().subscribe(
+        this.lookupService.getApiLookupOffences<any>().subscribe(
           (res) => {
             this.lookupData.offences = res.value;
             if (this.lookupData.offences) {
